requirements.txt
scikit-learn
pillow
numpy
streamlit
opencv-python-headless